fix(desktop): polish organization invitation members UI
